Retrospective changes
Some minor changes to variable names and comments. I was going to add
an option to replace duplicate files based on the hash making this
commit a bit more meaningful, but it would've added more statements to
the code than I would've liked.

This also removes the $max_size option, which should always just be
whatever your server/php.ini allows, so no reason to configure it
specifically in the script.

Also removes the sharex thing, which might be out of date now and is so
simple to configure anyway it was unnecessary.